refactor(userspace/sysdig/utils): clean up plugin utils and add methods for managing sources
